requirements.txt
Flask==2.3.3
python-dotenv
azure-core
azure-search-documents==11.5.2
azure-storage-blob
azure-identity
openai
aiohttp
ipywidgets
ipykernel
psycopg2==2.9.10
python3-saml==1.16.0
pyJWT
lxml==5.3.1
xmlsec==1.3.14
